Figured out that the mongodb schema should probably live in the class object (ie: Park.js). This would allow for each class to be linked by a different data source and referenced by the resolver
